requirements.txt
torch>=1.9.0
torch-geometric>=2.0.0
numpy>=1.19.0
pandas>=1.3.0
PyYAML>=5.4.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0